pyproject.toml
[tool.poetry]
name = "airqualityforecast"
version = "0.1.0"
description = ""
authors = ["Zsombor Teremy <teremyz21@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "3.10.10"
isort = "^5.13.2"
pandas = "^2.0.0"
requests = "^2.31.0"
pydantic = "^2.5.3"
hopsworks = "^3.0.5"
python-dotenv = "^1.0.0"
pyyaml = "^6.0.1"
python-box = "^7.1.1"
pathlib = "^1.0.1"
typer = "^0.9.0"
great-expectations = "0.14.13"
hsfs = "3.4.7"
urllib3 = "1.26.18"
botocore = "1.34.23"
types-requests = "2.25"
types-pyyaml = "^6.0.12.12"

[tool.poetry.group.dev.dependencies]
isort = "^5.13.2"
pytest-cov = "^4.1.0"
pre-commit = "^3.6.0"
flake8 = "^7.0.0"
mypy = "^1.8.0"
black = {version = "^24.1a1", allow-prereleases = true}
ipykernel = "^6.29.0"
pathlib = "^1.0.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.isort]
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
line_length = 79

[tool.black]
line-length = 79
target-version = ['py38']
include = '\.pyi?$'
exclude = '''

(
  /(
      \.eggs         # exclude a few common directories in the
    | \.git          # root of the project
    | \.hg
    | \.mypy_cache
    | \.tox
    | \.venv
    | _build
    | buck-out
    | build
    | dist
  )/
  | foo.py           # also separately exclude a file named foo.py in
                     # the root of the project
)
'''
